+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fix `AddBorutusAmountCounterCache` migration
+
+### Added
+- Add correct migration for `Borutus::Account.amounts_count`
+
 ## [0.2.2] - 2019-10-25
 ### Added
 - Add `Borutus::Account.with_amounts` scope that fetches accounts with amount entries